Introduce gpustat package and restructure related codes
